refactor/Remove redunant occurenies of function getDbConnectionParameters
